FEATURE: consolidates scenarios data/types into one file
